Made db access async

+++ src/main/java/Nick3306/github/io/NickQuest/QuestPlayer.java
@@ -41,7 +41,7 @@ public class QuestPlayer
 	{
 		this.exp = exp;
 	}
-	public int getSkillLevel(String skill)
+	public double getSkillLevel(String skill)
 	{
 		return skills.get(skill);
 	}
